lua: move serializer helpers into its own file
It is easier to glance on tightly coupled structures and functions, when
they're not mixed with others.

Just move without actual changes.

Part of #3228

(cherry picked from commit 1955cb1a0af089ba45c3493e65305889682fc01a)

A tiny difference from master's patch: moved msgpack.h include from
utils.h to lua/serializer.h and lua/msgpack.h.
