+++ src/lua/utils.c
@@ -39,8 +39,6 @@
 #include <fiber.h>
 
 int luaL_nil_ref = LUA_REFNIL;
-int luaL_map_metatable_ref = LUA_REFNIL;
-int luaL_array_metatable_ref = LUA_REFNIL;
 static int CTID_CONST_STRUCT_ERROR_REF = 0;
 
 static int luaT_newthread_ref = LUA_NOREF;
@@ -198,516 +196,6 @@ luaL_setcdatagc(struct lua_State *L, int idx)
 
 	/* Pop finalizer */
 	lua_pop(L, 1);
-}
-
-
-#define OPTION(type, name, defvalue) { #name, \
-	offsetof(struct luaL_serializer, name), type, defvalue}
-/**
- * Configuration options for serializers
- * @sa struct luaL_serializer
- */
-static struct {
-	const char *name;
-	size_t offset; /* offset in structure */
-	int type;
-	int defvalue;
-} OPTIONS[] = {
-	OPTION(LUA_TBOOLEAN, encode_sparse_convert, 1),
-	OPTION(LUA_TNUMBER,  encode_sparse_ratio, 2),
-	OPTION(LUA_TNUMBER,  encode_sparse_safe, 10),
-	OPTION(LUA_TNUMBER,  encode_max_depth, 32),
-	OPTION(LUA_TBOOLEAN, encode_deep_as_nil, 0),
-	OPTION(LUA_TBOOLEAN, encode_invalid_numbers, 1),
-	OPTION(LUA_TNUMBER,  encode_number_precision, 14),
-	OPTION(LUA_TBOOLEAN, encode_load_metatables, 1),
-	OPTION(LUA_TBOOLEAN, encode_use_tostring, 0),
-	OPTION(LUA_TBOOLEAN, encode_invalid_as_nil, 0),
-	OPTION(LUA_TBOOLEAN, decode_invalid_numbers, 1),
-	OPTION(LUA_TBOOLEAN, decode_save_metatables, 1),
-	OPTION(LUA_TNUMBER,  decode_max_depth, 32),
-	{ NULL, 0, 0, 0},
-};
-
-void
-luaL_serializer_create(struct luaL_serializer *cfg)
-{
-	rlist_create(&cfg->on_update);
-	for (int i = 0; OPTIONS[i].name != NULL; i++) {
-		int *pval = (int *) ((char *) cfg + OPTIONS[i].offset);
-		*pval = OPTIONS[i].defvalue;
-	}
-}
-
-void
-luaL_serializer_copy_options(struct luaL_serializer *dst,
-			     const struct luaL_serializer *src)
-{
-	memcpy(dst, src, offsetof(struct luaL_serializer, end_of_options));
-}
-
-/**
- * Configure one field in @a cfg. Value of the field is kept on
- * Lua stack after this function, and should be popped manually.
- * @param L Lua stack.
- * @param i Index of option in OPTIONS[].
- * @param cfg Serializer to inherit configuration.
- * @retval Pointer to the value of option.
- * @retval NULL if option is not in the table.
- */
-static int *
-luaL_serializer_parse_option(struct lua_State *L, int i,
-			     struct luaL_serializer *cfg)
-{
-	lua_getfield(L, 2, OPTIONS[i].name);
-	if (lua_isnil(L, -1))
-		return NULL;
-	/*
-	 * Update struct luaL_serializer using pointer to a
-	 * configuration value (all values must be `int` for that).
-	*/
-	int *pval = (int *) ((char *) cfg + OPTIONS[i].offset);
-	switch (OPTIONS[i].type) {
-	case LUA_TBOOLEAN:
-		*pval = lua_toboolean(L, -1);
-		break;
-	case LUA_TNUMBER:
-		*pval = lua_tointeger(L, -1);
-		break;
-	default:
-		unreachable();
-	}
-	return pval;
-}
-
-void
-luaL_serializer_parse_options(struct lua_State *L,
-			      struct luaL_serializer *cfg)
-{
-	for (int i = 0; OPTIONS[i].name != NULL; ++i) {
-		luaL_serializer_parse_option(L, i, cfg);
-		lua_pop(L, 1);
-	}
-}
-
-/**
- * @brief serializer.cfg{} Lua binding for serializers.
- * serializer.cfg is a table that contains current configuration values from
- * luaL_serializer structure. serializer.cfg has overriden __call() method
- * to change configuration keys in internal userdata (like box.cfg{}).
- * Please note that direct change in serializer.cfg.key will not affect
- * internal state of userdata. Changes via cfg() are reflected in
- * both Lua cfg table, and C serializer structure.
- * @param L lua stack
- * @return 0
- */
-static int
-luaL_serializer_cfg(struct lua_State *L)
-{
-	/* Serializer.cfg */
-	luaL_checktype(L, 1, LUA_TTABLE);
-	/* Updated parameters. */
-	luaL_checktype(L, 2, LUA_TTABLE);
-	struct luaL_serializer *cfg = luaL_checkserializer(L);
-	for (int i = 0; OPTIONS[i].name != NULL; ++i) {
-		if (luaL_serializer_parse_option(L, i, cfg) == NULL)
-			lua_pop(L, 1);
-		else
-			lua_setfield(L, 1, OPTIONS[i].name);
-	}
-	trigger_run(&cfg->on_update, cfg);
-	return 0;
-}
-
-/**
- * @brief serializer.new() Lua binding.
- * @param L stack
- * @param reg methods to register
- * @param parent parent serializer to inherit configuration
- * @return new serializer
- */
-struct luaL_serializer *
-luaL_newserializer(struct lua_State *L, const char *modname, const luaL_Reg *reg)
-{
-	luaL_checkstack(L, 1, "too many upvalues");
-
-	/* Create new module */
-	lua_newtable(L);
-
-	/* Create new configuration */
-	struct luaL_serializer *serializer = (struct luaL_serializer *)
-			lua_newuserdata(L, sizeof(*serializer));
-	luaL_getmetatable(L, LUAL_SERIALIZER);
-	lua_setmetatable(L, -2);
-	luaL_serializer_create(serializer);
-
-	for (; reg->name != NULL; reg++) {
-		/* push luaL_serializer as upvalue */
-		lua_pushvalue(L, -1);
-		/* register method */
-		lua_pushcclosure(L, reg->func, 1);
-		lua_setfield(L, -3, reg->name);
-	}
-
-	/* Add cfg{} */
-	lua_newtable(L); /* cfg */
-	lua_newtable(L); /* metatable */
-	lua_pushvalue(L, -3); /* luaL_serializer */
-	lua_pushcclosure(L, luaL_serializer_cfg, 1);
-	lua_setfield(L, -2, "__call");
-	lua_setmetatable(L, -2);
-	/* Save configuration values to serializer.cfg */
-	for (int i = 0; OPTIONS[i].name != NULL; i++) {
-		int *pval = (int *) ((char *) serializer + OPTIONS[i].offset);
-		switch (OPTIONS[i].type) {
-		case LUA_TBOOLEAN:
-			lua_pushboolean(L, *pval);
-			break;
-		case LUA_TNUMBER:
-			lua_pushinteger(L, *pval);
-			break;
-		default:
-			unreachable();
-		}
-		lua_setfield(L, -2, OPTIONS[i].name);
-	}
-	lua_setfield(L, -3, "cfg");
-
-	lua_pop(L, 1);  /* remove upvalues */
-
-	luaL_pushnull(L);
-	lua_setfield(L, -2, "NULL");
-	lua_rawgeti(L, LUA_REGISTRYINDEX, luaL_array_metatable_ref);
-	lua_setfield(L, -2, "array_mt");
-	lua_rawgeti(L, LUA_REGISTRYINDEX, luaL_map_metatable_ref);
-	lua_setfield(L, -2, "map_mt");
-
-	if (modname != NULL) {
-		/* Register module */
-		lua_getfield(L, LUA_REGISTRYINDEX, "_LOADED");
-		lua_pushstring(L, modname); /* add alias */
-		lua_pushvalue(L, -3);
-		lua_settable(L, -3);
-		lua_pop(L, 1); /* _LOADED */
-	}
-
-	return serializer;
-}
-
-static int
-lua_gettable_wrapper(lua_State *L)
-{
-	lua_gettable(L, -2);
-	return 1;
-}
-
-static void
-lua_field_inspect_ucdata(struct lua_State *L, struct luaL_serializer *cfg,
-			int idx, struct luaL_field *field)
-{
-	if (!cfg->encode_load_metatables)
-		return;
-
-	/*
-	 * Try to call LUAL_SERIALIZE method on udata/cdata
-	 * LuaJIT specific: lua_getfield/lua_gettable raises exception on
-	 * cdata if field doesn't exist.
-	 */
-	int top = lua_gettop(L);
-	lua_pushcfunction(L, lua_gettable_wrapper);
-	lua_pushvalue(L, idx);
-	lua_pushliteral(L, LUAL_SERIALIZE);
-	if (lua_pcall(L, 2, 1, 0) == 0  && !lua_isnil(L, -1)) {
-		if (!lua_isfunction(L, -1))
-			luaL_error(L, "invalid " LUAL_SERIALIZE  " value");
-		/* copy object itself */
-		lua_pushvalue(L, idx);
-		lua_pcall(L, 1, 1, 0);
-		/* replace obj with the unpacked value */
-		lua_replace(L, idx);
-		luaL_tofield(L, cfg, idx, field);
-	} /* else ignore lua_gettable exceptions */
-	lua_settop(L, top); /* remove temporary objects */
-}
-
-static void
-lua_field_inspect_table(struct lua_State *L, struct luaL_serializer *cfg,
-			int idx, struct luaL_field *field)
-{
-	assert(lua_type(L, idx) == LUA_TTABLE);
-	const char *type;
-	uint32_t size = 0;
-	uint32_t max = 0;
-
-	/* Try to get field LUAL_SERIALIZER_TYPE from metatable */
-	if (!cfg->encode_load_metatables ||
-	    !luaL_getmetafield(L, idx, LUAL_SERIALIZE))
-		goto skip;
-
-	if (lua_isfunction(L, -1)) {
-		/* copy object itself */
-		lua_pushvalue(L, idx);
-		lua_call(L, 1, 1);
-		/* replace obj with the unpacked value */
-		lua_replace(L, idx);
-		luaL_tofield(L, cfg, idx, field);
-		return;
-	} else if (!lua_isstring(L, -1)) {
-		luaL_error(L, "invalid " LUAL_SERIALIZE  " value");
-	}
-
-	type = lua_tostring(L, -1);
-	if (strcmp(type, "array") == 0 || strcmp(type, "seq") == 0 ||
-	    strcmp(type, "sequence") == 0) {
-		field->type = MP_ARRAY; /* Override type */
-		field->size = luaL_arrlen(L, idx);
-		/* YAML: use flow mode if __serialize == 'seq' */
-		if (cfg->has_compact && type[3] == '\0')
-			field->compact = true;
-		lua_pop(L, 1); /* type */
-
-		return;
-	} else if (strcmp(type, "map") == 0 || strcmp(type, "mapping") == 0) {
-		field->type = MP_MAP;   /* Override type */
-		field->size = luaL_maplen(L, idx);
-		/* YAML: use flow mode if __serialize == 'map' */
-		if (cfg->has_compact && type[3] == '\0')
-			field->compact = true;
-		lua_pop(L, 1); /* type */
-		return;
-	} else {
-		luaL_error(L, "invalid " LUAL_SERIALIZE "  value");
-	}
-
-skip:
-	field->type = MP_ARRAY;
-
-	/* Calculate size and check that table can represent an array */
-	lua_pushnil(L);
-	while (lua_next(L, idx)) {
-		size++;
-		lua_pop(L, 1); /* pop the value */
-		lua_Number k;
-		if (lua_type(L, -1) != LUA_TNUMBER ||
-		    ((k = lua_tonumber(L, -1)) != size &&
-		     (k < 1 || floor(k) != k))) {
-			/* Finish size calculation */
-			while (lua_next(L, idx)) {
-				size++;
-				lua_pop(L, 1); /* pop the value */
-			}
-			field->type = MP_MAP;
-			field->size = size;
-			return;
-		}
-		if (k > max)
-			max = k;
-	}
-
-	/* Encode excessively sparse arrays as objects (if enabled) */
-	if (cfg->encode_sparse_ratio > 0 &&
-	    max > size * (uint32_t)cfg->encode_sparse_ratio &&
-	    max > (uint32_t)cfg->encode_sparse_safe) {
-		if (!cfg->encode_sparse_convert)
-			luaL_error(L, "excessively sparse array");
-		field->type = MP_MAP;
-		field->size = size;
-		return;
-	}
-
-	assert(field->type == MP_ARRAY);
-	field->size = max;
-}
-
-static void
-lua_field_tostring(struct lua_State *L, struct luaL_serializer *cfg, int idx,
-		   struct luaL_field *field)
-{
-	int top = lua_gettop(L);
-	lua_getglobal(L, "tostring");
-	lua_pushvalue(L, idx);
-	lua_call(L, 1, 1);
-	lua_replace(L, idx);
-	lua_settop(L, top);
-	luaL_tofield(L, cfg, idx, field);
-}
-
-void
-luaL_tofield(struct lua_State *L, struct luaL_serializer *cfg, int index,
-		 struct luaL_field *field)
-{
-	if (index < 0)
-		index = lua_gettop(L) + index + 1;
-
-	double num;
-	double intpart;
-	size_t size;
-
-#define CHECK_NUMBER(x) ({\
-	if (!isfinite(x) && !cfg->encode_invalid_numbers) {		\
-		if (!cfg->encode_invalid_as_nil)				\
-			luaL_error(L, "number must not be NaN or Inf");		\
-		field->type = MP_NIL;						\
-	}})
-
-	switch (lua_type(L, index)) {
-	case LUA_TNUMBER:
-		num = lua_tonumber(L, index);
-		if (isfinite(num) && modf(num, &intpart) != 0.0) {
-			field->type = MP_DOUBLE;
-			field->dval = num;
-		} else if (num >= 0 && num < exp2(64)) {
-			field->type = MP_UINT;
-			field->ival = (uint64_t) num;
-		} else if (num >= -exp2(63) && num < exp2(63)) {
-			field->type = MP_INT;
-			field->ival = (int64_t) num;
-		} else {
-			field->type = MP_DOUBLE;
-			field->dval = num;
-			CHECK_NUMBER(num);
-		}
-		return;
-	case LUA_TCDATA:
-	{
-		uint32_t ctypeid = 0;
-		void *cdata = luaL_checkcdata(L, index, &ctypeid);
-		int64_t ival;
-		switch (ctypeid) {
-		case CTID_BOOL:
-			field->type = MP_BOOL;
-			field->bval = *(bool*) cdata;
-			return;
-		case CTID_CCHAR:
-		case CTID_INT8:
-			ival = *(int8_t *) cdata;
-			field->type = (ival >= 0) ? MP_UINT : MP_INT;
-			field->ival = ival;
-			return;
-		case CTID_INT16:
-			ival = *(int16_t *) cdata;
-			field->type = (ival >= 0) ? MP_UINT : MP_INT;
-			field->ival = ival;
-			return;
-		case CTID_INT32:
-			ival = *(int32_t *) cdata;
-			field->type = (ival >= 0) ? MP_UINT : MP_INT;
-			field->ival = ival;
-			return;
-		case CTID_INT64:
-			ival = *(int64_t *) cdata;
-			field->type = (ival >= 0) ? MP_UINT : MP_INT;
-			field->ival = ival;
-			return;
-		case CTID_UINT8:
-			field->type = MP_UINT;
-			field->ival = *(uint8_t *) cdata;
-			return;
-		case CTID_UINT16:
-			field->type = MP_UINT;
-			field->ival = *(uint16_t *) cdata;
-			return;
-		case CTID_UINT32:
-			field->type = MP_UINT;
-			field->ival = *(uint32_t *) cdata;
-			return;
-		case CTID_UINT64:
-			field->type = MP_UINT;
-			field->ival = *(uint64_t *) cdata;
-			return;
-		case CTID_FLOAT:
-			field->type = MP_FLOAT;
-			field->fval = *(float *) cdata;
-			CHECK_NUMBER(field->fval);
-			return;
-		case CTID_DOUBLE:
-			field->type = MP_DOUBLE;
-			field->dval = *(double *) cdata;
-			CHECK_NUMBER(field->dval);
-			return;
-		case CTID_P_CVOID:
-		case CTID_P_VOID:
-			if (*(void **) cdata == NULL) {
-				field->type = MP_NIL;
-				return;
-			}
-			/* Fall through */
-		default:
-			field->type = MP_EXT;
-			return;
-		}
-		return;
-	}
-	case LUA_TBOOLEAN:
-		field->type = MP_BOOL;
-		field->bval = lua_toboolean(L, index);
-		return;
-	case LUA_TNIL:
-		field->type = MP_NIL;
-		return;
-	case LUA_TSTRING:
-		field->sval.data = lua_tolstring(L, index, &size);
-		field->sval.len = (uint32_t) size;
-		field->type = MP_STR;
-		return;
-	case LUA_TTABLE:
-	{
-		field->compact = false;
-		lua_field_inspect_table(L, cfg, index, field);
-		return;
-	}
-	case LUA_TLIGHTUSERDATA:
-	case LUA_TUSERDATA:
-		field->sval.data = NULL;
-		field->sval.len = 0;
-		if (lua_touserdata(L, index) == NULL) {
-			field->type = MP_NIL;
-			return;
-		}
-		/* Fall through */
-	default:
-		field->type = MP_EXT;
-		return;
-	}
-#undef CHECK_NUMBER
-}
-
-void
-luaL_convertfield(struct lua_State *L, struct luaL_serializer *cfg, int idx,
-		  struct luaL_field *field)
-{
-	if (idx < 0)
-		idx = lua_gettop(L) + idx + 1;
-	assert(field->type == MP_EXT); /* must be called after tofield() */
-
-	if (cfg->encode_load_metatables) {
-		int type = lua_type(L, idx);
-		if (type == LUA_TCDATA) {
-			/*
-			 * Don't call __serialize on primitive types
-			 * https://github.com/tarantool/tarantool/issues/1226
-			 */
-			GCcdata *cd = cdataV(L->base + idx - 1);
-			if (cd->ctypeid > CTID_CTYPEID)
-				lua_field_inspect_ucdata(L, cfg, idx, field);
-		} else if (type == LUA_TUSERDATA) {
-			lua_field_inspect_ucdata(L, cfg, idx, field);
-		}
-	}
-
-	if (field->type == MP_EXT && cfg->encode_use_tostring)
-		lua_field_tostring(L, cfg, idx, field);
-
-	if (field->type != MP_EXT)
-		return;
-
-	if (cfg->encode_invalid_as_nil) {
-		field->type = MP_NIL;
-		return;
-	}
-
-	luaL_error(L, "unsupported Lua type '%s'",
-		   lua_typename(L, lua_type(L, idx)));
 }
 
 /**
@@ -1137,10 +625,6 @@ luaT_newthread(struct lua_State *L)
 int
 tarantool_lua_utils_init(struct lua_State *L)
 {
-	static const struct luaL_Reg serializermeta[] = {
-		{NULL, NULL},
-	};
-
 	/* Get CTypeID for `struct error *' */
 	int rc = luaL_cdef(L, "struct error;");
 	assert(rc == 0);
@@ -1148,26 +632,9 @@ tarantool_lua_utils_init(struct lua_State *L)
 	CTID_CONST_STRUCT_ERROR_REF = luaL_ctypeid(L, "const struct error &");
 	assert(CTID_CONST_STRUCT_ERROR_REF != 0);
 
-	luaL_register_type(L, LUAL_SERIALIZER, serializermeta);
 	/* Create NULL constant */
 	*(void **) luaL_pushcdata(L, CTID_P_VOID) = NULL;
 	luaL_nil_ref = luaL_ref(L, LUA_REGISTRYINDEX);
-
-	lua_createtable(L, 0, 1);
-	lua_pushliteral(L, "map"); /* YAML will use flow mode */
-	lua_setfield(L, -2, LUAL_SERIALIZE);
-	/* automatically reset hints on table change */
-	luaL_loadstring(L, "setmetatable((...), nil); return rawset(...)");
-	lua_setfield(L, -2, "__newindex");
-	luaL_map_metatable_ref = luaL_ref(L, LUA_REGISTRYINDEX);
-
-	lua_createtable(L, 0, 1);
-	lua_pushliteral(L, "seq"); /* YAML will use flow mode */
-	lua_setfield(L, -2, LUAL_SERIALIZE);
-	/* automatically reset hints on table change */
-	luaL_loadstring(L, "setmetatable((...), nil); return rawset(...)");
-	lua_setfield(L, -2, "__newindex");
-	luaL_array_metatable_ref = luaL_ref(L, LUA_REGISTRYINDEX);
 
 	rc = luaL_cdef(L, "struct ibuf;");
 	assert(rc == 0);
